+++ hemi/contrib/import.go
@@ -10,6 +10,8 @@ package contrib
 import (
 	_ "github.com/hexinfra/gorox/hemi/contrib/addons/demo"
 	_ "github.com/hexinfra/gorox/hemi/contrib/backends/redis"
+	_ "github.com/hexinfra/gorox/hemi/contrib/cachers/local"
+	_ "github.com/hexinfra/gorox/hemi/contrib/cachers/mem"
 	_ "github.com/hexinfra/gorox/hemi/contrib/cronjobs/clean"
 	_ "github.com/hexinfra/gorox/hemi/contrib/cronjobs/stat"
 	_ "github.com/hexinfra/gorox/hemi/contrib/filters/tcps/echo"
@@ -31,6 +33,4 @@ import (
 	_ "github.com/hexinfra/gorox/hemi/contrib/servers/socks"
 	_ "github.com/hexinfra/gorox/hemi/contrib/socklets/hello"
 	_ "github.com/hexinfra/gorox/hemi/contrib/staters/local"
-	_ "github.com/hexinfra/gorox/hemi/contrib/storers/local"
-	_ "github.com/hexinfra/gorox/hemi/contrib/storers/mem"
 )