+++ hemi/internal/component.go
@@ -33,7 +33,7 @@ const ( // component list
 	compUDPSFilter            // udpsProxy, ...
 	compCase                  // case
 	compStater                // localStater, redisStater, ...
-	compStorer                // localStorer, redisStorer, ...
+	compCacher                // localCacher, redisCacher, ...
 	compApp                   // app
 	compHandlet               // static, ...
 	compReviser               // gzipReviser, wrapReviser, ...
@@ -80,7 +80,7 @@ var (
 	tcpsFilterCreators = make(map[string]func(name string, stage *Stage, mesher *TCPSMesher) TCPSFilter)
 	udpsFilterCreators = make(map[string]func(name string, stage *Stage, mesher *UDPSMesher) UDPSFilter)
 	staterCreators     = make(map[string]func(name string, stage *Stage) Stater)
-	storerCreators     = make(map[string]func(name string, stage *Stage) Storer)
+	cacherCreators     = make(map[string]func(name string, stage *Stage) Cacher)
 	handletCreators    = make(map[string]func(name string, stage *Stage, app *App) Handlet)
 	reviserCreators    = make(map[string]func(name string, stage *Stage, app *App) Reviser)
 	sockletCreators    = make(map[string]func(name string, stage *Stage, app *App) Socklet)
@@ -106,8 +106,8 @@ func RegisterUDPSFilter(sign string, create func(name string, stage *Stage, mesh
 func RegisterStater(sign string, create func(name string, stage *Stage) Stater) {
 	_registerComponent0(sign, compStater, staterCreators, create)
 }
-func RegisterStorer(sign string, create func(name string, stage *Stage) Storer) {
-	_registerComponent0(sign, compStorer, storerCreators, create)
+func RegisterCacher(sign string, create func(name string, stage *Stage) Cacher) {
+	_registerComponent0(sign, compCacher, cacherCreators, create)
 }
 func RegisterHandlet(sign string, create func(name string, stage *Stage, app *App) Handlet) {
 	_registerComponent1(sign, compHandlet, handletCreators, create)
@@ -125,7 +125,7 @@ func RegisterCronjob(sign string, create func(name string, stage *Stage) Cronjob
 	_registerComponent0(sign, compCronjob, cronjobCreators, create)
 }
 
-func _registerComponent0[T Component](sign string, comp int16, creators map[string]func(string, *Stage) T, create func(string, *Stage) T) { // addon, backend, stater, storer, server, cronjob
+func _registerComponent0[T Component](sign string, comp int16, creators map[string]func(string, *Stage) T, create func(string, *Stage) T) { // addon, backend, stater, cacher, server, cronjob
 	creatorsLock.Lock()
 	defer creatorsLock.Unlock()
 
@@ -361,7 +361,7 @@ type Stage struct {
 	tcpsMeshers  compDict[*TCPSMesher] // indexed by mesherName
 	udpsMeshers  compDict[*UDPSMesher] // indexed by mesherName
 	staters      compDict[Stater]      // indexed by staterName
-	storers      compDict[Storer]      // indexed by storerName
+	cachers      compDict[Cacher]      // indexed by cacherName
 	apps         compDict[*App]        // indexed by appName
 	svcs         compDict[*Svc]        // indexed by svcName
 	servers      compDict[Server]      // indexed by serverName
@@ -416,7 +416,7 @@ func (s *Stage) onCreate() {
 	s.tcpsMeshers = make(compDict[*TCPSMesher])
 	s.udpsMeshers = make(compDict[*UDPSMesher])
 	s.staters = make(compDict[Stater])
-	s.storers = make(compDict[Storer])
+	s.cachers = make(compDict[Cacher])
 	s.apps = make(compDict[*App])
 	s.svcs = make(compDict[*Svc])
 	s.servers = make(compDict[Server])
@@ -443,9 +443,9 @@ func (s *Stage) OnShutdown() {
 	s.apps.goWalk((*App).OnShutdown)
 	s.WaitSubs()
 
-	// storers & staters
-	s.IncSub(len(s.storers) + len(s.staters))
-	s.storers.goWalk(Storer.OnShutdown)
+	// cachers & staters
+	s.IncSub(len(s.cachers) + len(s.staters))
+	s.cachers.goWalk(Cacher.OnShutdown)
 	s.staters.goWalk(Stater.OnShutdown)
 	s.WaitSubs()
 
@@ -550,7 +550,7 @@ func (s *Stage) OnConfigure() {
 	s.tcpsMeshers.walk((*TCPSMesher).OnConfigure)
 	s.udpsMeshers.walk((*UDPSMesher).OnConfigure)
 	s.staters.walk(Stater.OnConfigure)
-	s.storers.walk(Storer.OnConfigure)
+	s.cachers.walk(Cacher.OnConfigure)
 	s.apps.walk((*App).OnConfigure)
 	s.svcs.walk((*Svc).OnConfigure)
 	s.servers.walk(Server.OnConfigure)
@@ -571,7 +571,7 @@ func (s *Stage) OnPrepare() {
 	s.tcpsMeshers.walk((*TCPSMesher).OnPrepare)
 	s.udpsMeshers.walk((*UDPSMesher).OnPrepare)
 	s.staters.walk(Stater.OnPrepare)
-	s.storers.walk(Storer.OnPrepare)
+	s.cachers.walk(Cacher.OnPrepare)
 	s.apps.walk((*App).OnPrepare)
 	s.svcs.walk((*Svc).OnPrepare)
 	s.servers.walk(Server.OnPrepare)
@@ -647,18 +647,18 @@ func (s *Stage) createStater(sign string, name string) Stater {
 	s.staters[name] = stater
 	return stater
 }
-func (s *Stage) createStorer(sign string, name string) Storer {
-	if s.Storer(name) != nil {
-		UseExitf("conflicting storer with a same name '%s'\n", name)
+func (s *Stage) createCacher(sign string, name string) Cacher {
+	if s.Cacher(name) != nil {
+		UseExitf("conflicting cacher with a same name '%s'\n", name)
 	}
-	create, ok := storerCreators[sign]
+	create, ok := cacherCreators[sign]
 	if !ok {
-		UseExitln("unknown storer type: " + sign)
+		UseExitln("unknown cacher type: " + sign)
 	}
-	storer := create(name, s)
-	storer.setShell(storer)
-	s.storers[name] = storer
-	return storer
+	cacher := create(name, s)
+	cacher.setShell(cacher)
+	s.cachers[name] = cacher
+	return cacher
 }
 func (s *Stage) createApp(name string) *App {
 	if s.App(name) != nil {
@@ -729,7 +729,7 @@ func (s *Stage) QUICMesher(name string) *QUICMesher { return s.quicMeshers[name]
 func (s *Stage) TCPSMesher(name string) *TCPSMesher { return s.tcpsMeshers[name] }
 func (s *Stage) UDPSMesher(name string) *UDPSMesher { return s.udpsMeshers[name] }
 func (s *Stage) Stater(name string) Stater          { return s.staters[name] }
-func (s *Stage) Storer(name string) Storer          { return s.storers[name] }
+func (s *Stage) Cacher(name string) Cacher          { return s.cachers[name] }
 func (s *Stage) App(name string) *App               { return s.apps[name] }
 func (s *Stage) Svc(name string) *Svc               { return s.svcs[name] }
 func (s *Stage) Server(name string) Server          { return s.servers[name] }
@@ -785,7 +785,7 @@ func (s *Stage) Start(id int32) {
 	s.startBackends() // go backend.maintain()
 	s.startMeshers()  // go mesher.serve()
 	s.startStaters()  // go stater.Maintain()
-	s.startStorers()  // go storer.Maintain()
+	s.startCachers()  // go cacher.Maintain()
 	s.startApps()     // go app.maintain()
 	s.startSvcs()     // go svc.maintain()
 	s.startServers()  // go server.Serve()
@@ -871,12 +871,12 @@ func (s *Stage) startStaters() {
 		go stater.Maintain()
 	}
 }
-func (s *Stage) startStorers() {
-	for _, storer := range s.storers {
+func (s *Stage) startCachers() {
+	for _, cacher := range s.cachers {
 		if Debug() >= 1 {
-			Printf("storer=%s go Maintain()\n", storer.Name())
+			Printf("cacher=%s go Maintain()\n", cacher.Name())
 		}
-		go storer.Maintain()
+		go cacher.Maintain()
 	}
 }
 func (s *Stage) startApps() {