+++ HISTORY.md
@@ -3,6 +3,7 @@ v?.?.?                                                        (2023-??-?? UTC+8)
 
   * Interface of TCPSFilter is refactored.
   * Runner is renamed as Addon.
+  * Storers are renamed as cachers.
 
 v0.1.7                                                        (2023-06-13 UTC+8)
 ================================================================================